+++ src/module/config.js
@@ -90,7 +90,6 @@ export const OSE = {
   },
   monster_saves: {
     0: {
-      label: "Normal Human",
       d: 14,
       w: 15,
       p: 16,
@@ -98,7 +97,6 @@ export const OSE = {
       s: 18
     },
     1: {
-      label: "1-3",
       d: 12,
       w: 13,
       p: 14,
@@ -106,7 +104,6 @@ export const OSE = {
       s: 16
     },
     4: {
-      label: "4-6",
       d: 10,
       w: 11,
       p: 12,
@@ -114,7 +111,6 @@ export const OSE = {
       s: 14
     },
     7: {
-      label: "7-9",
       d: 8,
       w: 9,
       p: 10,
@@ -122,7 +118,6 @@ export const OSE = {
       s: 12
     },
     10: {
-      label: "10-12",
       d: 6,
       w: 7,
       p: 8,
@@ -130,7 +125,6 @@ export const OSE = {
       s: 10
     },
     13: {
-      label: "13-15",
       d: 4,
       w: 5,
       p: 6,
@@ -138,7 +132,6 @@ export const OSE = {
       s: 8
     },
     16: {
-      label: "16-18",
       d: 2,
       w: 3,
       p: 4,
@@ -146,7 +139,6 @@ export const OSE = {
       s: 6
     },
     19: {
-      label: "19-21",
       d: 2,
       w: 2,
       p: 2,
@@ -154,12 +146,29 @@ export const OSE = {
       s: 4
     },
     22: {
-      label: "22+",
       d: 2,
       w: 2,
       p: 2,
       b: 2,
       s: 2
     },
+  },
+  monster_thac0: {
+    0: 20,
+    1: 19,
+    2: 18,
+    3: 17,
+    4: 16,
+    5: 15,
+    6: 14,
+    7: 13,
+    9: 12,
+    10: 11,
+    12: 10,
+    14: 9,
+    16: 8,
+    18: 7,
+    20: 6,
+    22: 5
   }
 };